+++ backend/src/app.module.ts
@@ -4,12 +4,14 @@ import { AppService } from './app.service';
 import { PrismaModule } from './common/prisma.module';
 import { UsersModule } from './modules/users/users.module';
 import { AuthModule } from './modules/auth/auth.module';
+import { WorkspacesModule } from './modules/workspaces/workspaces.module';
 
 @Module({
   imports: [
     PrismaModule, // Подключаем PrismaModule - теперь PrismaService доступен везде
     UsersModule, // Подключаем модуль пользователей
     AuthModule, // Подключаем модуль аутентификации
+    WorkspacesModule, // Подключаем модуль workspace
   ],
   controllers: [AppController],
   providers: [AppService],